# docs/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  handlers.JWTOutput:
    properties:
      expires:
        type: string
      token:
        type: string
    type: object
  models.Company:
    description: Company
    properties:
      address:
        description: Company Address
        type: string
      contacts:
        description: Contacts
        items:
          $ref: '#/definitions/models.Contact'
        type: array
      description:
        description: Some important details
        type: string
      id:
        type: string
      info:
        description: Company Info
        type: string
      name:
        description: Company Name
        type: string
      publishedAt:
        type: string
      site:
        description: Company Site
        type: string
      updatedAt:
        description: Date of last update
        type: string
    type: object
  models.Contact:
    description: Company Contacts
    properties:
      email:
        description: Email
        type: string
      firstName:
        description: First name
        type: string
      info:
        description: Info
        type: string
      lastName:
        description: Last name
        type: string
      phone:
        description: Phone/cel
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: mentesan@gmail.com
    name: Fabio Almeida
    url: https://github.com/mentesan
  license:
    name: MIT
    url: https://github.com/mentesan/pert-api/blob/main/LICENSE
  title: Pentest Execution and Reporting Tool (PERT) API
  version: 0.0.1
paths:
  /companies:
    delete:
      parameters:
      - description: Company.ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Delete company
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Company'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get companies list
    post:
      parameters:
      - description: JSON for new company
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Company'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      summary: Add new company
    put:
      parameters:
      - description: Company.ID
        in: query
        name: id
        required: true
        type: string
      - description: All fields are optional
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Company'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      summary: Update company information
  /companies/search:
    get:
      parameters:
      - description: For search
        in: body
        name: request
        required: true
        schema:
          example: '{ "name": "CompanyName" }'
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Search company by Name
...
  /refresh:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.JWTOutput'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      summary: Refresh token
  /signin:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.JWTOutput'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      summary: Sign In User
  /signout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      summary: Sign out user
...
produces:
- application/json
schemes:
- http
swagger: "2.0"